Enhancement: Simplify wallet connect flow by defaulting to CLN (#250)
